Add unix makefile

+++ include/Common.h
@@ -5,6 +5,8 @@
 extern "C" {
 #endif
 
+#include <limits.h>
+
 #define RB_OK ( 0 )
 #define RB_TRUE ( 1 )
 #define RB_FALSE ( 0 )
@@ -16,7 +18,7 @@ extern "C" {
 #define RB_INVALID_ARG ( RB_ERROR_BASE - 3 )
 #define RB_TIMEOUT ( RB_ERROR_BASE - 4 )
 
-#define RB_WAIT_INFINITE ( LONG_MIN )
+#define RB_WAIT_INFINITE ( INT_MIN )
 
 #ifdef __cplusplus
 }
